Update OSPREY.py v.1.0.0 -> v.1.1.0
Additional functionality has been added to OSPREY. v.1.1.0 is able to export a summary PDF file, containing a score breakdown and total score for the student. 

A patch has also been added to fix a bug in the significant figures check that used to force integers to be answered as, say, 5.0 rather than 5. 

Two additional questions were also added to the end of the smart worksheet.